Move devices code
Move it down one level and have a ‘devices’ function that enables/disables the devices based on config

We set the NODE_PATH to make it nicer for the device requires
