+++ cli/devices/camera/camera-settings.js
@@ -1,11 +1,12 @@
 const debug = require( 'debug' )( 'biab:camera:settings' );
 const fs = require( 'fs' );
-const config = require( '../config' );
+const config = require( 'config' );
 
 function saveSettings( wp, settings ) {
 	debug( 'Camera settings: ' + settings );
 
 	config.set( 'camera_settings', { args: settings } );
+	this.emit( 'result', 'ok' );
 }
 
 module.exports = saveSettings;
