+++ cli/devices/camera/constants.js
