+++ cli/devices/camera/photo-to-post.js
