+++ cli/devices/sensehat/capture.js
