+++ cli/devices/sensehat/capture.py
