+++ cli/devices/sensehat/display/clear.py
