+++ cli/devices/sensehat/display/show-image.py
