+++ cli/devices/sensehat/display/show-message.py
