+++ cli/devices/sensehat/index.js
@@ -1,9 +1,13 @@
+const debug = require( 'debug' )( 'biab:sensehat' );
+
 const schedule = require( './schedule' );
 const dataToWP = require( './data-to-wp' );
 const capture = require( './capture' );
 const display = require( './display' );
 
 module.exports = function( emitter ) {
+	debug( 'Enabling Sensehat' );
+
 	emitter.on( 'sensehat-capture', capture );
 	emitter.on( 'sensehat-reading', dataToWP );
 	emitter.on( 'sensehat-schedule', schedule );
